refactor: config 패키지 내에 관심사 별로 하위 패키지로 분리
