+++ src/main/java/roomescape/config/auth/AuthWebMvcConfig.java
@@ -1,12 +1,10 @@
-package roomescape.config;
+package roomescape.config.auth;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.controller.auth.LoginMemberArgumentResolver;
-import roomescape.controller.auth.RoleAllowedInterceptor;
 import roomescape.controller.login.CookieExtractor;
 import roomescape.service.login.LoginService;
