+++ src/main/java/roomescape/config/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.controller.auth;
+package roomescape.config.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
