+++ src/main/java/roomescape/config/auth/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.controller.auth;
+package roomescape.config.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
