+++ src/main/java/roomescape/config/auth/RoleAllowedInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.controller.auth;
+package roomescape.config.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
