+++ src/main/java/roomescape/controller/member/MemberController.java
@@ -3,7 +3,7 @@ package roomescape.controller.member;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.auth.RoleAllowed;
+import roomescape.config.auth.RoleAllowed;
 import roomescape.domain.member.MemberRole;
 import roomescape.service.member.MemberService;
 import roomescape.service.member.dto.MemberListResponse;
