+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.auth.LoginMember;
-import roomescape.controller.auth.RoleAllowed;
+import roomescape.config.auth.LoginMember;
+import roomescape.config.auth.RoleAllowed;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRole;
 import roomescape.service.member.MemberService;
