+++ src/main/java/roomescape/controller/reservationtime/ReservationTimeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.auth.RoleAllowed;
+import roomescape.config.auth.RoleAllowed;
 import roomescape.domain.member.MemberRole;
 import roomescape.service.reservationtime.ReservationTimeService;
 import roomescape.service.reservationtime.dto.ReservationTimeAvailableListResponse;
