+++ src/main/java/roomescape/controller/view/AdminPageController.java
@@ -3,7 +3,7 @@ package roomescape.controller.view;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.controller.auth.RoleAllowed;
+import roomescape.config.auth.RoleAllowed;
 import roomescape.domain.member.MemberRole;
 
 @RequestMapping("/admin")
